increased id column size

+++ docker/schema.sql
@@ -1,6 +1,6 @@
 CREATE TABLE public.whale_alerts
 (
-    id               int4         NOT NULL,
+    id               int8         NOT NULL,
     message          text         NULL,
     link             varchar(255) NULL,
     posted_at        timestamp    NULL,
@@ -19,7 +19,7 @@ CREATE INDEX "IDX_PROCESSED_STATUS" ON public.whale_alerts USING btree (process_
 
 CREATE TABLE public.alert_details
 (
-    alert_id        int4         NOT NULL,
+    alert_id        int8         NOT NULL,
     blockchain      varchar(32)  NULL,
     "type"          varchar(32)  NULL,
     hash            varchar(128) NULL,
